'新闻发布判断'

+++ messagecenter-biz/src/main/java/org/smartwork/biz/service/impl/ZGNewsServiceImpl.java
@@ -4,8 +4,11 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.CollectionUtils;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.baomidou.mybatisplus.extension.toolkit.SqlHelper;
+import org.forbes.comm.exception.ForbesException;
 import org.forbes.comm.utils.ConvertUtils;
 import org.smartwork.biz.service.IZGNewsService;
+import org.smartwork.comm.enums.MsgBizResultEnum;
+import org.smartwork.comm.enums.ReleaseStateEnum;
 import org.smartwork.dal.entity.ZGNews;
 import org.smartwork.dal.entity.ZGNewsAttach;
 import org.smartwork.dal.mapper.ZGNewsAttachMapper;
@@ -43,6 +46,10 @@ public class ZGNewsServiceImpl extends ServiceImpl<ZGNewsMapper, ZGNews> impleme
     @Override
     @Transactional(rollbackFor = {Exception.class})
     public boolean updateById(ZGNews news) {
+        ZGNews oldNews = baseMapper.selectById(news.getId());
+        if(ReleaseStateEnum.HAVE_RELEASED.getCode().equals(oldNews.getReleaseState())){
+            throw new ForbesException(MsgBizResultEnum.HAVE_RELEASED.getBizCode(),MsgBizResultEnum.HAVE_RELEASED.getBizMessage());
+        }
         boolean isUpdate =  SqlHelper.retBool(baseMapper.updateById(news));
         Long newsId = news.getId();
         newsAttachMapper.delete(new QueryWrapper<ZGNewsAttach>().eq("news_id",newsId));
@@ -60,6 +67,10 @@ public class ZGNewsServiceImpl extends ServiceImpl<ZGNewsMapper, ZGNews> impleme
     @Override
     @Transactional(rollbackFor = {Exception.class})
     public boolean removeById(Serializable id) {
+        ZGNews oldNews = baseMapper.selectById(id);
+        if(ReleaseStateEnum.HAVE_RELEASED.getCode().equals(oldNews.getReleaseState())){
+            throw new ForbesException(MsgBizResultEnum.HAVE_RELEASED_DEL.getBizCode(),MsgBizResultEnum.HAVE_RELEASED_DEL.getBizMessage());
+        }
         boolean isDel =  SqlHelper.retBool(baseMapper.deleteById(id));
         newsAttachMapper.delete(new QueryWrapper<ZGNewsAttach>().eq("news_id",id));
         return isDel;
@@ -68,8 +79,30 @@ public class ZGNewsServiceImpl extends ServiceImpl<ZGNewsMapper, ZGNews> impleme
     @Override
     @Transactional(rollbackFor = {Exception.class})
     public boolean removeByIds(Collection<? extends Serializable> idList) {
+        Integer exisCount = baseMapper.selectCount(new QueryWrapper<ZGNews>().in("id",idList)
+                .eq("release_state",ReleaseStateEnum.HAVE_RELEASED.getCode()));
+        if(exisCount >0 ){
+            throw new ForbesException(MsgBizResultEnum.HAVE_RELEASED_DEL.getBizCode(),MsgBizResultEnum.HAVE_RELEASED_DEL.getBizMessage());
+        }
         boolean isDel =  CollectionUtils.isEmpty(idList) ? false : SqlHelper.retBool(baseMapper.deleteBatchIds(idList));
         newsAttachMapper.delete(new QueryWrapper<ZGNewsAttach>().in("news_id",idList));
         return isDel;
+    }
+
+
+    /***
+     *
+     * @param id
+     * @return
+     */
+    @Override
+    public  ZGNews getById(Serializable id)
+    {
+        ZGNews news = baseMapper.selectById(id);
+        List<ZGNewsAttach> newsAttachs = newsAttachMapper.selectList(new QueryWrapper<ZGNewsAttach>().eq("news_id",id));
+        if(ConvertUtils.isNotEmpty(newsAttachs)){
+            news.setNewsAttachs(newsAttachs);
+        }
+        return news;
     }
 }
